moves pieces out of components to model
